test(refactor): move 'fundWalletIdFromLightning', add 'updatePending' check
